Added Repository Class to store data
